requirements.txt
google-api-python-client
Flask
biopython
requests
synbiochem-py